Add top-level CMakeLists
Add top-level CMakeLists.txt. It introduces new build system based
on components/libraries of structure compliant with the new file structure.
CmakeLists.txt is used by build.sh to build the application, tools,
and tests. When used, it also generates TestSuitesList.txt in the build
directory. TestSuitesList.txt is intended to be used for running tests
with run_test.py script which will be added later.